Implemented Zombie AI that chases you and bites when close enough.

+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 //to recieve damage.
 public interface IDamageable
 {
-    public void TakeDamage(int damage, GameObject other);
+    public void TakeDamage(float damage, GameObject other);
 }
